Full Commit of all required scripts for the pi
3 Main scripts include
CycleEDM.py - cycles power to the EDM if the relay is connected

BOOTSTRAP.py - can flash a new program file to the EDM

loggingGit.py - Main python script for the automatic logging system. Helper functions can be found in the src folder.

The src folder also contains several .txt files used for unit testing for the helper functions. (I havent figured out how to move these tests out of the scripts they relate to so all of that is still in this folder)